#4157 - remove element-content - not needed
